wait for webchat framework hook
