+++ src/applications/virtual-agent/containers/App.jsx
@@ -1,8 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import WaitForFeatureToggles from './WaitForFeatureToggles';
 import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
 import ChatbotError from './ChatbotError';
+import useWaitForWebchatFramework from './useWaitForWebchatFramework';
 
 window.React = React;
 window.ReactDOM = ReactDOM;
@@ -18,51 +19,6 @@ const loadWebChat = () => {
 };
 
 loadWebChat();
-
-function checkForWebchat(
-  setLoading,
-  setError,
-  MAX_INTERVAL_CALL_COUNT,
-  timeout,
-) {
-  let intervalCallCount = 0;
-  const intervalId = setInterval(() => {
-    intervalCallCount++;
-    if (window.WebChat) {
-      setLoading(false);
-      setError(false);
-      clearInterval(intervalId);
-    } else if (intervalCallCount > MAX_INTERVAL_CALL_COUNT) {
-      setError(true);
-      setLoading(false);
-      clearInterval(intervalId);
-    }
-  }, timeout);
-}
-
-function useWaitForWebchatFramework(props) {
-  const [isLoading, setLoading] = useState(!window.WebChat);
-  const [error, setError] = useState(false);
-
-  const TIMEOUT_DURATION_MS = 250;
-  const DEFAULT_WEBCHAT_TIMEOUT = 1 * 60 * 1000;
-
-  const webchatTimeout = props.webchatTimeout
-    ? props.webchatTimeout
-    : DEFAULT_WEBCHAT_TIMEOUT;
-  const MAX_INTERVAL_CALL_COUNT = webchatTimeout / TIMEOUT_DURATION_MS;
-
-  if (isLoading) {
-    checkForWebchat(
-      setLoading,
-      setError,
-      MAX_INTERVAL_CALL_COUNT,
-      TIMEOUT_DURATION_MS,
-    );
-  }
-
-  return { isLoading, error };
-}
 
 export default function App(props) {
   const { isLoading, error } = useWaitForWebchatFramework(props);
